refactor(guard): extract runtime artifact policy to keep oversized files from growing
